requirements.txt
pydantic==2.12.5
pydantic-settings==2.12.0
pyyaml==6.0.3
python-dotenv==1.0.1
google-cloud-firestore==2.17.2
google-cloud-aiplatform==1.128.0
google-generativeai==0.5.4
google-genai==1.53.0
google-adk==1.20.0
google-cloud-logging==3.12.1
google-cloud-trace==1.16.1
requests==2.32.5
beautifulsoup4==4.12.3
pypdf==4.3.1
youtube-transcript-api==0.6.2
pytest==8.3.2
pytest-mock==3.14.0
fastapi==0.118.3
uvicorn==0.38.0
starlette==0.48.0
